fix: retrofit response

+++ app/src/main/java/com/example/kinopoisk/mainFragment/models/Rating.kt
@@ -1,5 +1,5 @@
 package com.example.kinopoisk.mainFragment.models
 
 data class Rating(
-    val imdb: Int
+    val imdb: Double
 )
